+++ app/models/rimrock_computation.rb
@@ -5,12 +5,6 @@ class RimrockComputation < Computation
   validates :input_path, :output_path, :working_file_name, absence: true
   validates :revision, presence: true, unless: :created?
 
-  # TODO: this method should be remove after
-  # https://gitlab.com/eurvalve/vapor/issues/237 is resolved
-  def revision
-    'master'
-  end
-
   private
 
   def created?