+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Gitlab integration services (list branches/tags and download specific file) (@amber7b)
 - Fetch computation slurm start script from Gitlab (@mkasztelnik)
 - Pipeline comparison view uses OFF viewers to show 3D mesh differences (@dharezlak)
+- Enabled selection of rimrock computation version (@jmeizner)
 
 ### Changed
 - Upgrade to ruby 2.4.1 (@mkasztelnik)